refactor(weddings): remove duplicate overview endpoint in favor of reporting dashboard
- Remove duplicate GET /api/v1/weddings/{uuid}/overview/ endpoint and WeddingOverviewOut schema.
- Update WeddingDetailPage and useWeddingDetail on frontend to directly consume useDashboardWedding.
- Declare dashboard schemas strictly in apps.reporting.schemas.
- Decouple apps.weddings entirely from cross-domain aggregations (zero foreign model queries in weddings).
- Regenerate OpenAPI spec and Orval client hooks.
